src/main/java/com/kakao/cafe/security/UserDetailsServiceImpl.java
package com.kakao.cafe.security;

import com.kakao.cafe.domain.User;
import com.kakao.cafe.error.exception.nonexist.UserNotFoundedException;
import com.kakao.cafe.service.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;

@RequiredArgsConstructor
public class UserDetailsServiceImpl implements UserDetailsService {

    private final UserService userService;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user;
        try {
            user = userService.findByEmail(username);
        } catch (UserNotFoundedException e) {
            throw new UsernameNotFoundException("User not found with: " + username);
        }
        return UserDetailsVO.of(user);
    }
}
